* info/man.c (xrefs_of_manpage): use int for consistency.
* info/display.c (display_process_line), info/util.c (ansi_escape)
(printed_representation), info/man.c (clean_manpage), info/window.c
(calculate_line_starts, window_compute_line_map): use int for
ansi_escape len and printed_representation pbytes and pchars for
consistency and also in case wcwidth in printed_representation returns
a negative value.

+++ info/display.c
@@ -330,8 +330,8 @@ display_process_line (WINDOW *win,
 {
   mbi_iterator_t iter;
   const char *cur_ptr;
-  size_t pchars = 0; /* Printed chars */
-  size_t pbytes = 0; /* Bytes to output. */
+  int pchars = 0; /* Printed chars */
+  int pbytes = 0; /* Bytes to output. */
   char *rep;
   int in_match = 0;
   int in_ref = 0, in_ref_proper = 0;
